fix copy button js
the issue wasn't with the location of the js I guess, but permissions...

at console when deployed, when copy button was clicked the following was dispalyed:
macros/:91 Refused to execute inline event handler because it violates the following Content Security Policy directive: "script-src 'self' talk.hyvor.com". Either the 'unsafe-inline' keyword, a hash ('sha256-...'), or a nonce ('nonce-...') is required to enable inline execution. Note that hashes do not apply to event handlers, style attributes and javascript: navigations unless the 'unsafe-hashes' keyword is present.

this is a workaround to test