bugfix tab enter nav custom
Prior behaviour:
If I type in "randomAllergen" with no match or autosuggestions, the only
way to move to the diameter input is to ENTER.
If I type If I type in "randomAllergen" with no match or
autosuggestions, click TAB (nothing happens), and then press ENTER,
nothing occurs. I no longer move to the diameter input.
If I type If I type in "randomAllergen" with no match or
autosuggestions, click TAB (nothing happens), and then press ENTER,
nothing occurs. If I press TAB again, I now am able to move to the
diameter input.

Now: If I type If I type in "randomAllergen" with no match or
autosuggestions, click TAB => I move to the diameter input.